refactor: Point와 ValidatePoint 영역 분리
